fix(audit): installable 1.0 assets, clippy clean, Rust CI, Go boundary, main protection
- Upload v1.0.0 Release binaries matching SHA256SUMS (35 artifacts + metadata)
- scripts/verify-release-assets.sh + science-go-module-boundary.sh
- cargo clippy -D warnings clean for xai-grok-science; 262 lib tests pass
- Science CI runs cargo test+clippy and Go product-scope boundary
- Honest 5.0 WP evidence levels; 1.0 status notes downloadable assets
- Protect main: required status checks, no force-push/delete

+++ agent/crates/codegen/xai-grok-science/src/seqbench.rs
@@ -388,12 +388,12 @@ fn genetic_code(codon: &str) -> char {
 fn find_orfs(seq: &str, rna: bool, min_aa: usize) -> Vec<Orf> {
     let mut out = Vec::new();
     for frame in 1..=3 {
-        out.extend(orfs_on_strand(seq, frame as i32, min_aa));
+        out.extend(orfs_on_strand(seq, frame, min_aa));
     }
     let rc = reverse_complement(seq, rna);
     for frame in 1..=3 {
-        for mut o in orfs_on_strand(&rc, frame as i32, min_aa) {
-            o.frame = -(frame as i32);
+        for mut o in orfs_on_strand(&rc, frame, min_aa) {
+            o.frame = -frame;
             out.push(o);
         }
     }
